+++ R/reexports.R
@@ -1,0 +1,12 @@
+#' @importFrom insight display
+#' @export
+insight::display
+
+#' @importFrom insight print_md
+#' @export
+insight::print_md
+
+
+#' @importFrom insight print_html
+#' @export
+insight::print_html